requirements.txt
asgiref==3.8.1
certifi==2025.1.31
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.1
cloudinary==1.43.0
dj3-cloudinary-storage==0.0.6
Django==5.1.7
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
idna==3.10
msgpack==1.1.0
pillow==11.1.0
PyJWT==2.9.0
redis==5.2.1
requests==2.32.3
rest-framework-simplejwt==0.0.2
six==1.17.0
sqlparse==0.5.3
tzdata==2025.1
urllib3==2.3.0
